Refactored rtk.datasets and have fixed the diffusion pipeline and metric generation (maybe?). Merging.
