+++ rtk/utils.py
@@ -8,6 +8,7 @@ import os
 import pandas as pd
 import yaml
 import hydra
+from copy import deepcopy
 from argparse import Namespace
 from colorlog import ColoredFormatter
 from logging import Logger
